test(shared): consolidate duplicate config, utils, and app test suites (#27171)
