+++ packages/shared/src/config/distribution-profile.test.ts
@@ -4,6 +4,7 @@
 import { describe, expect, it } from "vitest";
 
 import {
+  DISTRIBUTION_PROFILES,
   isDistributionProfile,
   resolveDistributionProfile,
 } from "./distribution-profile.js";
@@ -25,6 +26,10 @@ describe("isDistributionProfile", () => {
     expect(isDistributionProfile(null)).toBe(false);
     expect(isDistributionProfile(42)).toBe(false);
     expect(isDistributionProfile({})).toBe(false);
+  });
+
+  it("exposes the canonical profile list", () => {
+    expect([...DISTRIBUTION_PROFILES]).toEqual(["store", "unrestricted"]);
   });
 });
