+++ packages/shared/src/dev-settings-figlet-heading.test.ts
@@ -12,7 +12,29 @@ import {
   renderDevSubsystemFigletHeading,
 } from "./dev-settings-figlet-heading.js";
 
-describe("dev-settings-figlet-heading", () => {
+describe("renderDevSubsystemFigletHeading", () => {
+  it("renders a boxed marker with the subsystem text", () => {
+    const out = renderDevSubsystemFigletHeading("api");
+    expect(out).toContain("| API |");
+    expect(out.split("\n")).toHaveLength(3);
+  });
+
+  it("maps each subsystem kind to its text", () => {
+    expect(renderDevSubsystemFigletHeading("orchestrator")).toContain(
+      "ORCHESTRATOR",
+    );
+    expect(renderDevSubsystemFigletHeading("vite")).toContain("VITE");
+    expect(renderDevSubsystemFigletHeading("electrobun")).toContain(
+      "ELECTROBUN",
+    );
+  });
+
+  it("box width matches the text length", () => {
+    const out = renderDevSubsystemFigletHeading("vite");
+    const lines = out.split("\n");
+    expect(lines[0].length).toBe(lines[1].length);
+  });
+
   const kinds: DevSubsystemBannerKind[] = [
     "orchestrator",
     "vite",
@@ -29,7 +51,18 @@ describe("dev-settings-figlet-heading", () => {
 
     const lines = heading.split("\n");
     expect(lines.length).toBe(3);
+    expect(lines[0].length).toBe(lines[1].length);
     expect(lines[1]).toBe(`| ${kind.toUpperCase()} |`);
+  });
+});
+
+describe("prependDevSubsystemFigletHeading", () => {
+  it("separates heading from the table with a blank line", () => {
+    const out = prependDevSubsystemFigletHeading("api", "| table |");
+    const parts = out.split("\n\n");
+    expect(parts).toHaveLength(2);
+    expect(parts[0]).toContain("API");
+    expect(parts[1]).toBe("| table |");
   });
 
   it("prepends figlet heading above settings table with blank line separation", () => {
